feat: adds the CardBarChart component
- Creates a new component called CardBarChart
